Colorize Peek View to match theme
Change the Peek View defaults inherited from VSCode to match the overall
look and feel of Plastic. In the current state, Peek View looks out of
place. This change utilizes the two main background colors of the theme
to stylize the Peek View, but swaps them to add a visual distinction
from the main editor window.